Contact page bug with GetThisSemestersId
The GetThisSemestersId method now returns the current semester if
DateTime.Now falls within the defined dates for any stored semester.  If
no such semester exists (such as with breaks), it will return the
semester with the maximum StartDate.  Errors return null.  This presents
a couple known issues:
- Semesters, when created, should not overlap (validation)
- If we are in a semester, allow the creation of the next one, no more
and never BEFORE that point.  Until automated jobs are figured out, this
has to be manual.  If we are after a semester (someone hasn't added it
yet), allow the creation of the next one.  Look at the number of weeks
from the new StartDate to the most recent EndDate and verify that way.